Add files via upload
Write a program that inputs two integers n and k, where n=k. Your program should calculate the number of different ways that k bishops could be placed on an nXn chessboard.
Structure your program using the backtracking scheme that we have used for the eight queens
problem. What needs to be modified is the " OK " function.